changed percent multiplier to double to maybe make the bar more finegrained

+++ src/main/java/com/minecolonies/util/ExperienceUtils.java
@@ -13,7 +13,7 @@ public final class ExperienceUtils
     /**
      * The number to create a percentage from another number (ex. 100*0.25 = 25).
      */
-    private static final int PERCENT_MULTIPLIER = 100;
+    private static final double PERCENT_MULTIPLIER = 100D;
 
     /**
      * Private constructor to hide the public one.
